fix(swiftgit/windows): UInt32 cast on remaining libgit2 enum rawValues
Sweep through the rest of the SwiftGit lib for the same Apple/Linux
(UInt32 rawValue) vs Windows (Int32 rawValue, MSVC default underlying
type) mismatch on libgit2 bitfield enums. Touches:

  - GitClient+Init.swift: GIT_REPOSITORY_INIT_* flag composition.
  - GitClient+Log.swift: GIT_SORT_TIME | GIT_SORT_TOPOLOGICAL.
  - GitClient+Merge.swift / GitClient+Rebase.swift / GitClient.swift:
    GIT_CHECKOUT_SAFE / GIT_CHECKOUT_FORCE assigns to checkout_strategy.
  - GitClient+Stash.swift: GIT_STASH_*, GIT_STASH_APPLY_REINSTATE_INDEX.
  - GitClient+Status.swift: GIT_STATUS_OPT_* composition.

+++ Sources/SwiftGit/Lib/GitClient+Rebase.swift
@@ -164,7 +164,7 @@ extension GitClient {
             var coOpts = git_checkout_options()
             try check(git_checkout_options_init(
                 &coOpts, UInt32(GIT_CHECKOUT_OPTIONS_VERSION)))
-            coOpts.checkout_strategy = GIT_CHECKOUT_FORCE.rawValue
+            coOpts.checkout_strategy = UInt32(GIT_CHECKOUT_FORCE.rawValue)
             try check(git_checkout_head(repo, &coOpts))
 
             let total = Int(git_rebase_operation_entrycount(rebase))
